refactor: :art: break Search view into components
